+++ extensions/git-commands/src/utils.ts
@@ -2,16 +2,14 @@ import { Color, getPreferenceValues } from "@raycast/api";
 import { default as data } from "./alias.json";
 import { Alias, AliasType } from "./types";
 
-const preferences = getPreferenceValues();
-export const showDescription = preferences.ItemAliasShowDescription;
-export const showTypeColor = preferences.ItemAliasTypeColor;
+export const preferences = getPreferenceValues();
 
 export function getData() {
   return data as Alias[];
 }
 
 export function typeColor(type: AliasType) {
-  if (!showTypeColor) {
+  if (!preferences.ItemAliasTypeColor) {
     return Color.SecondaryText;
   }